Fixed wrong required size
git-svn-id: https://svn.codehaus.org/izpack/izpack-src/trunk@1897 7d736ef5-cfd4-0310-9c9a-b52d5c14b761

+++ src/lib/com/izforge/izpack/panels/PacksModel.java
@@ -489,7 +489,7 @@ class PacksModel extends AbstractTableModel
 
     private void updateBytes()
     {
-        int bytes = 0;
+        long bytes = 0;
         for (int q = 0; q < packs.size(); q++)
         {
             if (Math.abs(checkValues[q]) == 1)
